Remove old windowUI, rename class files
